backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-RESTX==0.5.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
python-dotenv==0.21.1
